+++ emis/settings.py
@@ -33,7 +33,6 @@ ALLOWED_HOSTS = []
 # Application definition
 
 INSTALLED_APPS = [
-    "sysadmin.apps.SysadminConfig",
     "web_app.apps.WebAppConfig",
     "django.contrib.admin",
     "django.contrib.auth",
