+++ web_app/forms/user.py
@@ -1,5 +1,5 @@
 from django import forms
-from django.contrib.auth.forms import UserCreationForm
+from django.contrib.auth.forms import UserCreationForm, UserChangeForm
 from django.contrib.auth.models import User
 from django.forms.models import ModelForm
 
@@ -18,3 +18,12 @@ class CustomUserCreationForm(ModelForm):
     class Meta:
         model = User
         fields = ("username", "first_name", "last_name", "email", "groups")
+
+class CustomEditUserForm(UserChangeForm):
+    first_name = forms.CharField(max_length=30, help_text="Required.")
+    last_name = forms.CharField(max_length=30, help_text="Required.")
+    password = None
+
+    class Meta:
+        model = User
+        fields = ("username", "first_name", "last_name", "email", "is_active", "groups")
