+++ web_app/models/activation.py
@@ -1,7 +1,6 @@
 from django.contrib.auth.models import User
 from django.db import models
 
-
 class Activation(models.Model):
     user = models.OneToOneField(
         User,
