+++ web_app/templates/web_app/base.html
@@ -45,8 +45,8 @@
                             System Admin
                         </a>
                         <div class="dropdown-menu" aria-labelledby="navbarDropdown">
-                            <a class="dropdown-item" href="{% url 'sysadmin:user-directory' %}">User Directory</a>
-                            <a class="dropdown-item" href="{% url 'sysadmin:create-user' %}">Create User</a>
+                            <a class="dropdown-item" href="{% url 'web_app:user-directory' %}">User Directory</a>
+                            <a class="dropdown-item" href="{% url 'web_app:create-user' %}">Create User</a>
                         </div>
                     </li>
                     {% endif %}
