+++ web_app/templates/web_app/sysadmin/user_activation_email.txt
